Integrating changes, still doesn't build

+++ rpi_version/rpi/build.sh
@@ -1,24 +1,77 @@
 #!/bin/bash
 
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+NC='\033[0m'
+
+base_directory=""
+obj_files_list=""
+
+function compile {
+  # Path to the current directory relative to the base directory.
+  local path=$1
+  
+  # Handle the subdirectories first.
+  for d in $base_directory/$path/*/ ; do
+    local base=$(basename "$d")
+    if [[ -d $d ]]; then
+      compile "$path/$base"
+    fi
+  done
+ 
+  # 1) Handle cc files.
+  for f in $base_directory/$path/*.cc ; do
+    local base=$(basename "$f")
+    if [[ -f $f ]]; then
+      local base_noext=$(echo $base | sed 's/\.[^.]*$//')
+      if [[ ! -d $base_directory/objs/$path ]]; then
+        mkdir -p $base_directory/objs/$path
+      fi
+      cd $base_directory/$path/
+      # g++ -I$base_directory/ -O3 -c $f -o $base_directory/objs/$path/$base_noext.o
+      g++ --std=c++0x -I$base_directory/ -g -c $f -o $base_directory/objs/$path/$base_noext.o -lwiringPi
+      if [[ $? -ne 0 ]]; then
+        echo -e "${RED}^ ${base}${NC}"
+        cd - > /dev/null
+        exit -1
+      else
+        echo -e "${GREEN}^ ${base}${NC}"
+        cd - > /dev/null
+        obj_files_list="$obj_files_list $base_directory/objs/$path/$base_noext.o"
+      fi
+    fi
+  done
+}
+
 rpi_hostname="vs-rpi"
 rpi_ip="192.168.1.254"
 rpi_user="pi"
 rpi_pass="12345"
 
 rpi_mtpt="/tmp/rpi_fs"
+rpi_code_dir="BB"
 
 if [ "$HOSTNAME" == "$rpi_hostname" ]; then
-  echo "remote"
+  echo "Running a remote build..."
+  cd $rpi_code_dir
+  base_directory=$PWD
+  compile "."
+  g++ --std=c++0x -g $obj_files_list -o $base_directory/run -lwiringPi
 else
   echo "Uploading code to the Raspberry Pi..."
   rm -rf $rpi_mtpt
   mkdir $rpi_mtpt
-  echo "yes" > /tmp/bla
-  echo "$rpi_pass" >> /tmp/bla
-  cat /tmp/bla | sshfs $rpi_user@$rpi_ip: $rpi_mtpt -C -o password_stdin
-  cp -r * "$rpi_mtpt/"
-  fusermount -u $rpi_mtpt
-  rm -rf $rpi_mtpt
+  echo "$rpi_pass" | sshfs $rpi_user@$rpi_ip: $rpi_mtpt -C -o password_stdin -o workaround=rename
+  if [ $? -eq 0 ]; then
+    rm -rf "$rpi_mtpt/$rpi_code_dir"
+    mkdir "$rpi_mtpt/$rpi_code_dir"
+    cp -r * "$rpi_mtpt/$rpi_code_dir/"
+    fusermount -u $rpi_mtpt
+    rm -rf $rpi_mtpt
+    echo "Sucessfully uploaded."
+  else
+    echo "Failed to upload!"
+    exit 1
+  fi
+  sshpass -p "$rpi_pass" ssh $rpi_user@$rpi_ip 'bash -s' < build.sh
 fi
-
-
